Common File Dialog の代わりに Common Item Dialog を使用
