+++ .gitattributes
@@ -1,1 +1,2 @@
 *.bat text eol=crlf
+Makefile text eolf=lf
