+++ sakura_core/String_define.h
@@ -1292,5 +1292,10 @@
 #define STR_ERR_FILEPATH_TOO_LONG		35034
 #define STR_GREP_EXCLUDE_FILE			35035
 #define STR_GREP_EXCLUDE_FOLDER			35036
+#define STR_FILEDIALOG_READONLY			35037
+#define STR_FILEDIALOG_CODE				35038
+#define STR_FILEDIALOG_EOL				35039
+#define STR_FILEDIALOG_MRU				35040
+#define STR_FILEDIALOG_OPENFOLDER		35041
 
-// Now using max number 35036 by STR_GREP_EXCLUDE_FOLDER
+// Now using max number 35041 by STR_FILEDIALOG_OPENFOLDER
