+++ sakura_core/sakura_rc.h
@@ -807,6 +807,7 @@
 #define IDC_STATIC_URL_GITHUB_PR        1733
 #define IDC_COMBO_EXCLUDE_FOLDER        1734
 #define IDC_COMBO_EXCLUDE_FILE          1735
+#define IDC_CHECK_VISTA_STYLE_FILEDIALOG	1737
 #define IDS_AUTHOR_PAGE                 4054
 #define IDS_ABOUT_DESCRIPTION           4056
 #define IDD_TYPELIST                    5000
@@ -849,7 +850,7 @@
 #ifndef APSTUDIO_READONLY_SYMBOLS
 #define _APS_NEXT_RESOURCE_VALUE        230
 #define _APS_NEXT_COMMAND_VALUE         4057
-#define _APS_NEXT_CONTROL_VALUE         1737
+#define _APS_NEXT_CONTROL_VALUE         1738
 #define _APS_NEXT_SYMED_VALUE           104
 #endif
 #endif
