+++ sakura_core/sakura_rc.rc
@@ -1243,9 +1243,10 @@ BEGIN
     EDITTEXT        IDC_EDIT_FILEOPENDIR, 161, 100, 111, 12, ES_AUTOHSCROLL
     PUSHBUTTON      "...", IDC_BUTTON_FILEOPENDIR, 273, 100, 9, 12
     GROUPBOX        "ファイルダイアログの初期位置", IDC_STATIC, 156, 41, 130, 79, WS_GROUP
-    CONTROL         "改行コードNEL, PS, LSを有効にする(&O)", IDC_CHECK_ENABLEEXTEOL, "Button", BS_AUTOCHECKBOX | WS_GROUP | WS_TABSTOP, 161, 138, 122, 10
-    CONTROL         "矩形選択移動で選択をロックする(&B)", IDC_CHECK_BOXSELECTLOCK, "Button", BS_AUTOCHECKBOX | WS_TABSTOP, 161, 151, 122, 10
-    GROUPBOX        "その他", IDC_STATIC, 156, 125, 131, 42, WS_GROUP
+    CONTROL         "Vistaスタイルのファイルダイアログ(&I)", IDC_CHECK_VISTA_STYLE_FILEDIALOG, "Button", BS_AUTOCHECKBOX | WS_GROUP | WS_TABSTOP, 161, 128, 125, 10
+    CONTROL         "改行コードNEL, PS, LSを有効にする(&O)", IDC_CHECK_ENABLEEXTEOL, "Button", BS_AUTOCHECKBOX | WS_GROUP | WS_TABSTOP, 161, 158, 122, 10
+    CONTROL         "矩形選択移動で選択をロックする(&B)", IDC_CHECK_BOXSELECTLOCK, "Button", BS_AUTOCHECKBOX | WS_TABSTOP, 161, 171, 122, 10
+    GROUPBOX        "その他", IDC_STATIC, 156, 145, 131, 42, WS_GROUP
 END
 
 IDD_PROP_FILE DIALOGEX 0, 0, 293, 240
@@ -4157,6 +4158,11 @@ GDI リソース\t残り  %d%%\n\n"
 	STR_FILETREE_REPLACE_PATH_TO	"置換先文字列を指定してください"
 
 	STR_ERR_FILEPATH_TOO_LONG		"%ls\nというファイルを開けません。\nファイルのパスが長すぎます。"
+	STR_FILEDIALOG_READONLY			"読み取り専用ファイルとして開く(&R)"
+	STR_FILEDIALOG_CODE				"文字コードセット(&C):"
+	STR_FILEDIALOG_EOL				"改行コード(&E):"
+	STR_FILEDIALOG_MRU				"最近のファイル(&F):"
+	STR_FILEDIALOG_OPENFOLDER		"最近のフォルダ(&D):"
 
 END
